refactor: extract StoryTypeComparator and JointMethodAugmentor from StoryApiConfigurationManager
Extract two delegates to reduce StoryApiConfigurationManager from 587 to 403 lines:

- StoryTypeComparator: package-private enum (was inner TypeComparator)
- JointMethodAugmentor: package-private final class with static augment()
  entry point (was JOINTED_MODEL_TYPE, getFieldMethodNameHint,
  addMethodsToType x2, augmentTypeIfNecessary body, getArgumentField)

StoryApiConfigurationManager retains its public API surface unchanged:
- getTypeComparator() delegates to StoryTypeComparator.SINGLETON
- augmentTypeIfNecessary() delegates to JointMethodAugmentor.augment()
- BIPED_RESOURCE_TYPE stays (subclass dependency)

Cleaned 9 orphaned imports from manager (Visibility, FieldTemplate,
Field, Maps, Logger, IdentifierNameGenerator, DynamicResource,
JointArrayId, JointId).

Closes #661
